refactoring
Consistently use 'pass'/'phase' instead of 'passes'/'phases'
