+++ dialects/mem/pass/rw/remem_elim.cpp
@@ -1,4 +1,4 @@
-#include "dialects/mem/passes/rw/remem_elim.h"
+#include "dialects/mem/pass/rw/remem_elim.h"
 
 #include "dialects/mem/mem.h"
